Add: UI Refactor and some test
