+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/components/BankAccountsView.kt
@@ -5,38 +5,21 @@ package app.cybrid.sdkandroid.components
 import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
-import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
 import androidx.activity.compose.ManagedActivityResultLauncher
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.constraintlayout.compose.ConstraintLayout
-import androidx.constraintlayout.compose.Dimension
 import app.cybrid.sdkandroid.R
+import app.cybrid.sdkandroid.components.bankAccounts.compose.BankAccountsView_Done
+import app.cybrid.sdkandroid.components.bankAccounts.compose.BankAccountsView_Error
+import app.cybrid.sdkandroid.components.bankAccounts.compose.BankAccountsView_Loading
+import app.cybrid.sdkandroid.components.bankAccounts.compose.BankAccountsView_Required
 import app.cybrid.sdkandroid.components.bankAccounts.view.BankAccountsViewModel
-import app.cybrid.sdkandroid.core.Constants
-import app.cybrid.sdkandroid.ui.Theme.robotoFont
-import com.plaid.link.OpenPlaidLink
 import com.plaid.link.configuration.LinkTokenConfiguration
 import com.plaid.link.linkTokenConfiguration
-import com.plaid.link.result.LinkExit
 import com.plaid.link.result.LinkResult
-import com.plaid.link.result.LinkSuccess
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -51,7 +34,7 @@ Component(context, attrs, defStyle) {
 
     private var currentState = mutableStateOf(BankAccountsViewState.LOADING)
 
-    var bankAccountViewModel: BankAccountsViewModel? = null
+    var bankAccountsViewModel: BankAccountsViewModel? = null
 
     init {
 
@@ -59,12 +42,12 @@ Component(context, attrs, defStyle) {
         this.composeView = findViewById(R.id.composeContent)
     }
 
-    fun setViewModel(bankAccountViewModel: BankAccountsViewModel) {
+    fun setViewModel(bankAccountsViewModel: BankAccountsViewModel) {
 
-        this.bankAccountViewModel = bankAccountViewModel
-        this.currentState = bankAccountViewModel.uiState
+        this.bankAccountsViewModel = bankAccountsViewModel
+        this.currentState = bankAccountsViewModel.uiState
         this.initComposeView()
-        GlobalScope.launch { bankAccountViewModel.createWorkflow() }
+        GlobalScope.launch { bankAccountsViewModel.createWorkflow() }
     }
 
     private fun initComposeView() {
@@ -73,7 +56,7 @@ Component(context, attrs, defStyle) {
             compose.setContent {
                 BankAccountsView(
                     currentState = this.currentState,
-                    viewModel = bankAccountViewModel
+                    viewModel = bankAccountsViewModel
                 )
             }
         }
@@ -119,323 +102,6 @@ fun BankAccountsView(
 
             BankAccountsView.BankAccountsViewState.ERROR -> {
                 BankAccountsView_Error()
-            }
-        }
-    }
-}
-
-@Composable
-fun BankAccountsView_Loading() {
-
-    Box(
-        modifier = Modifier
-            .height(120.dp)
-    ) {
-        Column(
-            modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Checking Bank Accounts",
-                fontFamily = robotoFont,
-                fontWeight = FontWeight.Normal,
-                fontSize = 17.sp,
-                color = colorResource(id = R.color.primary_color)
-            )
-            CircularProgressIndicator(
-                modifier = Modifier
-                    .padding(top = 16.dp),
-                color = colorResource(id = R.color.primary_color)
-            )
-        }
-    }
-}
-
-@Composable
-fun BankAccountsView_Required(viewModel: BankAccountsViewModel?) {
-
-    // -- Activity Result for Plaid
-    val getPlaidResult = rememberLauncherForActivityResult(OpenPlaidLink()) {
-        when (it) {
-            is LinkSuccess -> {
-
-                if (it.metadata.accounts.size == 1) {
-
-                    viewModel?.uiState?.value = BankAccountsView.BankAccountsViewState.LOADING
-                    GlobalScope.launch {
-                        viewModel?.createExternalBankAccount(
-                            publicToken = it.publicToken,
-                            account = it.metadata.accounts[0])
-                    }
-                } else {
-                    // -- Log multiple accounts or empty accounts
-                    viewModel?.uiState?.value = BankAccountsView.BankAccountsViewState.ERROR
-                }
-            }
-            is LinkExit -> {}
-        }
-    }
-
-    // -- Content
-    ConstraintLayout(
-        modifier = Modifier
-            .fillMaxSize()
-            .testTag(Constants.AccountsViewTestTags.List.id)
-    ) {
-
-        val (text, buttons) = createRefs()
-
-        Row(
-            modifier = Modifier.constrainAs(text) {
-                start.linkTo(parent.start, margin = 0.dp)
-                top.linkTo(parent.top, margin = 0.dp)
-                end.linkTo(parent.end, margin = 0.dp)
-                bottom.linkTo(parent.bottom, margin = 0.dp)
-            },
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = "No banks accounts connected.",
-                modifier = Modifier
-                    .padding(start = 10.dp),
-                fontFamily = robotoFont,
-                fontWeight = FontWeight.Medium,
-                fontSize = 19.sp,
-                lineHeight = 32.sp,
-                color = colorResource(id = R.color.black)
-            )
-        }
-        // -- Buttons
-        ConstraintLayout(
-            Modifier.constrainAs(buttons) {
-                start.linkTo(parent.start, margin = 10.dp)
-                end.linkTo(parent.end, margin = 10.dp)
-                bottom.linkTo(parent.bottom, margin = 20.dp)
-                width = Dimension.fillToConstraints
-                height = Dimension.value(50.dp)
-            }
-        ) {
-
-            val (continueButton) = createRefs()
-
-            // -- Continue Button
-            Button(
-                onClick = {
-                    BankAccountsView.openPlaid(
-                        plaidToken = viewModel?.latestWorkflow?.plaidLinkToken!!,
-                        getPlaidResult = getPlaidResult)
-                },
-                modifier = Modifier
-                    .constrainAs(continueButton) {
-                        start.linkTo(parent.start, margin = 10.dp)
-                        end.linkTo(parent.end, margin = 10.dp)
-                        top.linkTo(parent.top, margin = 0.dp)
-                        bottom.linkTo(parent.bottom, margin = 0.dp)
-                        width = Dimension.fillToConstraints
-                        height = Dimension.fillToConstraints
-                    },
-                shape = RoundedCornerShape(10.dp),
-                elevation = ButtonDefaults.elevation(
-                    defaultElevation = 4.dp,
-                    pressedElevation = 4.dp,
-                    disabledElevation = 0.dp
-                ),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = colorResource(id = R.color.primary_color),
-                    contentColor = Color.White
-                )
-            ) {
-                Text(
-                    text = "Add account",
-                    color = Color.White,
-                    fontFamily = robotoFont,
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 14.sp,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun BankAccountsView_Done() {
-
-    // -- Content
-    ConstraintLayout(
-        modifier = Modifier
-            .fillMaxSize()
-            .testTag(Constants.AccountsViewTestTags.List.id)
-    ) {
-
-        val (text, buttons) = createRefs()
-
-        Row(
-            modifier = Modifier.constrainAs(text) {
-                start.linkTo(parent.start, margin = 0.dp)
-                top.linkTo(parent.top, margin = 0.dp)
-                end.linkTo(parent.end, margin = 0.dp)
-                bottom.linkTo(parent.bottom, margin = 0.dp)
-            },
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.kyc_verified),
-                contentDescription = "",
-                modifier = Modifier
-                    .padding(top = 5.dp)
-                    .padding(0.dp)
-                    .size(26.dp),
-                contentScale = ContentScale.Fit
-            )
-            Text(
-                text = "Account added successfully.",
-                modifier = Modifier
-                    .padding(start = 10.dp),
-                fontFamily = robotoFont,
-                fontWeight = FontWeight.Medium,
-                fontSize = 19.sp,
-                lineHeight = 32.sp,
-                color = colorResource(id = R.color.black)
-            )
-        }
-        // -- Buttons
-        ConstraintLayout(
-            Modifier.constrainAs(buttons) {
-                start.linkTo(parent.start, margin = 10.dp)
-                end.linkTo(parent.end, margin = 10.dp)
-                bottom.linkTo(parent.bottom, margin = 20.dp)
-                width = Dimension.fillToConstraints
-                height = Dimension.value(50.dp)
-            }
-        ) {
-
-            val (doneButton) = createRefs()
-
-            // -- Continue Button
-            Button(
-                onClick = {},
-                modifier = Modifier
-                    .constrainAs(doneButton) {
-                        start.linkTo(parent.start, margin = 10.dp)
-                        end.linkTo(parent.end, margin = 10.dp)
-                        top.linkTo(parent.top, margin = 0.dp)
-                        bottom.linkTo(parent.bottom, margin = 0.dp)
-                        width = Dimension.fillToConstraints
-                        height = Dimension.fillToConstraints
-                    },
-                shape = RoundedCornerShape(10.dp),
-                elevation = ButtonDefaults.elevation(
-                    defaultElevation = 4.dp,
-                    pressedElevation = 4.dp,
-                    disabledElevation = 0.dp
-                ),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = colorResource(id = R.color.primary_color),
-                    contentColor = Color.White
-                )
-            ) {
-                Text(
-                    text = "Continue",
-                    color = Color.White,
-                    fontFamily = robotoFont,
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 14.sp,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun BankAccountsView_Error() {
-
-    // -- Vars
-    val onBackPressedDispatcher = LocalOnBackPressedDispatcherOwner.current?.onBackPressedDispatcher
-
-    // -- Content
-    ConstraintLayout(
-        modifier = Modifier
-            .fillMaxSize()
-            .testTag(Constants.AccountsViewTestTags.List.id)
-    ) {
-
-        val (text, buttons) = createRefs()
-
-        Row(
-            modifier = Modifier.constrainAs(text) {
-                start.linkTo(parent.start, margin = 0.dp)
-                top.linkTo(parent.top, margin = 0.dp)
-                end.linkTo(parent.end, margin = 0.dp)
-                bottom.linkTo(parent.bottom, margin = 0.dp)
-            },
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.kyc_error),
-                contentDescription = "",
-                modifier = Modifier
-                    .padding(top = 5.dp)
-                    .padding(0.dp)
-                    .size(26.dp),
-                contentScale = ContentScale.Fit
-            )
-            Text(
-                text = "An error has occurred, try again.",
-                modifier = Modifier
-                    .padding(start = 10.dp),
-                fontFamily = robotoFont,
-                fontWeight = FontWeight.Medium,
-                fontSize = 19.sp,
-                lineHeight = 32.sp,
-                color = colorResource(id = R.color.black)
-            )
-        }
-        // -- Buttons
-        ConstraintLayout(
-            Modifier.constrainAs(buttons) {
-                start.linkTo(parent.start, margin = 10.dp)
-                end.linkTo(parent.end, margin = 10.dp)
-                bottom.linkTo(parent.bottom, margin = 20.dp)
-                width = Dimension.fillToConstraints
-                height = Dimension.value(50.dp)
-            }
-        ) {
-
-            val (doneButton) = createRefs()
-
-            // -- Continue Button
-            Button(
-                onClick = {
-                    onBackPressedDispatcher?.onBackPressed()
-                },
-                modifier = Modifier
-                    .constrainAs(doneButton) {
-                        start.linkTo(parent.start, margin = 10.dp)
-                        end.linkTo(parent.end, margin = 10.dp)
-                        top.linkTo(parent.top, margin = 0.dp)
-                        bottom.linkTo(parent.bottom, margin = 0.dp)
-                        width = Dimension.fillToConstraints
-                        height = Dimension.fillToConstraints
-                    },
-                shape = RoundedCornerShape(10.dp),
-                elevation = ButtonDefaults.elevation(
-                    defaultElevation = 4.dp,
-                    pressedElevation = 4.dp,
-                    disabledElevation = 0.dp
-                ),
-                colors = ButtonDefaults.buttonColors(
-                    backgroundColor = colorResource(id = R.color.primary_color),
-                    contentColor = Color.White
-                )
-            ) {
-                Text(
-                    text = "Done",
-                    color = Color.White,
-                    fontFamily = robotoFont,
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 14.sp,
-                )
             }
         }
     }
