+++ SDKAndroid/src/main/res/values-en/strings.xml
@@ -63,4 +63,13 @@
     <string name="kyc_view_required_begin_button">Begin</string>
     <string name="kyc_view_required_done_button">Done</string>
 
+    <!-- Bank Accounts Component -->
+    <string name="bank_accounts_view_loading_text">Checking Bank Accounts</string>
+    <string name="bank_accounts_view_required_text">No banks accounts connected.</string>
+    <string name="bank_accounts_view_required_button">Add account</string>
+    <string name="bank_accounts_view_done_text">Account added successfully.</string>
+    <string name="bank_accounts_view_done_button">Continue</string>
+    <string name="bank_accounts_view_error_text">An error has occurred, try again.</string>
+    <string name="bank_accounts_view_error_button">Done</string>
+
 </resources>
